+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Configure Gradle JVM Test Suite Plugin. (#259)
 - Replace `SECURE_SESSION_NO_TEE_PROVIDER` with `SECURE_SESSION_NO_TEE_FRAMEWORK`
   in `TeeSessionGenerationError`. (#265)
+- Add `final` keyword in `EncryptionService`. (#268)
 
 ### Dependency Upgrades